+++ packages/salesforcedx-vscode-core/src/commands/util/preconditionCheckers.ts
@@ -4,7 +4,7 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { isNullOrUndefined } from '@salesforce/salesforcedx-utils-vscode/src/helpers';
+import { isNullOrUndefined } from '@salesforce/salesforcedx-utils-vscode/out/src/helpers';
 import { PreconditionChecker } from '@salesforce/salesforcedx-utils-vscode/src/types';
 import { workspace } from 'vscode';
 import { notificationService } from '../../notifications';
